requirements.txt
requests
azure-identity>=1.24.0
azure-core
